Updated mongocsharpdriver to 1.2 Mongo collection count now returns long

+++ MongoDB.Web/Providers/MongoDBProfileProvider.cs
@@ -26,13 +26,13 @@ namespace MongoDB.Web.Providers
                 query = Query.And(query, Query.EQ("IsAnonymous", authenticationOption == ProfileAuthenticationOption.Anonymous));
             }
 
-            return this.mongoCollection.Remove(query).DocumentsAffected;
+            return (int)this.mongoCollection.Remove(query).DocumentsAffected;
         }
 
         public override int DeleteProfiles(string[] usernames)
         {
             var query = Query.And(Query.EQ("ApplicationName", this.ApplicationName), Query.In("Username", new BsonArray(usernames)));
-            return this.mongoCollection.Remove(query).DocumentsAffected;
+            return ( int )this.mongoCollection.Remove( query ).DocumentsAffected;
         }
 
         public override int DeleteProfiles(ProfileInfoCollection profiles)
@@ -63,13 +63,13 @@ namespace MongoDB.Web.Providers
         public override int GetNumberOfInactiveProfiles(ProfileAuthenticationOption authenticationOption, DateTime userInactiveSinceDate)
         {
             var query = GetQuery(authenticationOption, null, userInactiveSinceDate);
-            return this.mongoCollection.Count(query);
+            return ( int )this.mongoCollection.Count( query );
         }
 
         public override SettingsPropertyValueCollection GetPropertyValues(SettingsContext context, SettingsPropertyCollection collection)
         {
             var settingsPropertyValueCollection = new SettingsPropertyValueCollection();
-            
+
             if (context == null || collection == null || collection.Count < 1)
             {
                 return settingsPropertyValueCollection;
@@ -141,7 +141,7 @@ namespace MongoDB.Web.Providers
                 {
                     continue;
                 }
-                
+
                 if (!isAuthenticated && !(bool)settingsPropertyValue.Property.Attributes["AllowAnonymous"])
                 {
                     continue;
@@ -180,7 +180,7 @@ namespace MongoDB.Web.Providers
         {
             var query = GetQuery(authenticationOption, usernameToMatch, userInactiveSinceDate);
 
-            totalRecords = this.mongoCollection.Count(query);
+            totalRecords = ( int )this.mongoCollection.Count( query );
 
             var profileInfoCollection = new ProfileInfoCollection();
 
@@ -195,7 +195,7 @@ namespace MongoDB.Web.Providers
         private IMongoQuery GetQuery(ProfileAuthenticationOption authenticationOption, string usernameToMatch, DateTime? userInactiveSinceDate)
         {
             var query = Query.EQ("ApplicationName", this.ApplicationName);
-            
+
             if (authenticationOption != ProfileAuthenticationOption.All)
             {
                 query = Query.And(query, Query.EQ("IsAnonymous", authenticationOption == ProfileAuthenticationOption.Anonymous));
